Move constructor and assignment traits into type_traits
