+++ src/utl/public/utl/type_traits/utl_is_explicitly_constructible.h
@@ -2,11 +2,11 @@
 
 #pragma once
 
-#include "utl/type_traits/constructor_traits/utl_is_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_implicitly_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_constructible.h"
 #include "utl/type_traits/utl_common.h"
 #include "utl/type_traits/utl_declval.h"
+#include "utl/type_traits/utl_is_constructible.h"
+#include "utl/type_traits/utl_is_implicitly_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_constructible.h"
 #include "utl/type_traits/utl_is_x_convertible.h"
 #include "utl/type_traits/utl_logical_traits.h"
