+++ src/utl/public/utl/type_traits/utl_is_implicitly_constructible.h
@@ -2,9 +2,10 @@
 
 #pragma once
 
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_constructible.h"
 #include "utl/type_traits/utl_common.h"
 #include "utl/type_traits/utl_declval.h"
+#include "utl/type_traits/utl_is_default_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_constructible.h"
 #include "utl/type_traits/utl_is_x_convertible.h"
 #include "utl/type_traits/utl_logical_traits.h"
