--- src/utl/public/utl/type_traits/utl_is_x_constructible.h
@@ -1,18 +1,0 @@
-// Copyright 2023-2024 Bryan Wong
-
-#pragma once
-
-#include "utl/type_traits/constructor_traits/utl_is_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_copy_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_default_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_explicitly_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_implicitly_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_move_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_copy_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_default_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_move_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_trivially_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_trivially_copy_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_trivially_default_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_trivially_move_constructible.h"
