+++ src/utl/public/utl/type_traits/utl_std_traits.h
@@ -15,20 +15,36 @@
 #include "utl/type_traits/utl_enable_if.h"
 #include "utl/type_traits/utl_extent.h"
 #include "utl/type_traits/utl_is_arithmetic.h"
+#include "utl/type_traits/utl_is_assignable.h"
 #include "utl/type_traits/utl_is_base_of.h"
 #include "utl/type_traits/utl_is_class.h"
 #include "utl/type_traits/utl_is_complete.h"
 #include "utl/type_traits/utl_is_compound.h"
 #include "utl/type_traits/utl_is_const.h"
+#include "utl/type_traits/utl_is_constructible.h"
+#include "utl/type_traits/utl_is_copy_assignable.h"
+#include "utl/type_traits/utl_is_copy_constructible.h"
+#include "utl/type_traits/utl_is_default_constructible.h"
 #include "utl/type_traits/utl_is_enum.h"
+#include "utl/type_traits/utl_is_explicitly_constructible.h"
 #include "utl/type_traits/utl_is_floating_point.h"
 #include "utl/type_traits/utl_is_function.h"
 #include "utl/type_traits/utl_is_fundamental.h"
+#include "utl/type_traits/utl_is_implicitly_constructible.h"
 #include "utl/type_traits/utl_is_integral.h"
 #include "utl/type_traits/utl_is_lvalue_reference.h"
 #include "utl/type_traits/utl_is_member_function_pointer.h"
 #include "utl/type_traits/utl_is_member_object_pointer.h"
 #include "utl/type_traits/utl_is_member_pointer.h"
+#include "utl/type_traits/utl_is_move_assignable.h"
+#include "utl/type_traits/utl_is_move_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_assignable.h"
+#include "utl/type_traits/utl_is_nothrow_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_copy_assignable.h"
+#include "utl/type_traits/utl_is_nothrow_copy_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_default_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_move_assignable.h"
+#include "utl/type_traits/utl_is_nothrow_move_constructible.h"
 #include "utl/type_traits/utl_is_null_pointer.h"
 #include "utl/type_traits/utl_is_object.h"
 #include "utl/type_traits/utl_is_pointer.h"
@@ -37,11 +53,17 @@
 #include "utl/type_traits/utl_is_same.h"
 #include "utl/type_traits/utl_is_scalar.h"
 #include "utl/type_traits/utl_is_signed.h"
+#include "utl/type_traits/utl_is_trivially_assignable.h"
+#include "utl/type_traits/utl_is_trivially_constructible.h"
+#include "utl/type_traits/utl_is_trivially_copy_assignable.h"
+#include "utl/type_traits/utl_is_trivially_copy_constructible.h"
+#include "utl/type_traits/utl_is_trivially_default_constructible.h"
+#include "utl/type_traits/utl_is_trivially_move_assignable.h"
+#include "utl/type_traits/utl_is_trivially_move_constructible.h"
 #include "utl/type_traits/utl_is_union.h"
 #include "utl/type_traits/utl_is_unsigned.h"
 #include "utl/type_traits/utl_is_void.h"
 #include "utl/type_traits/utl_is_volatile.h"
-#include "utl/type_traits/utl_is_x_constructible.h"
 #include "utl/type_traits/utl_is_x_convertible.h"
 #include "utl/type_traits/utl_logical_traits.h"
 #include "utl/type_traits/utl_modify_pointer.h"
