+++ src/utl/public/utl/type_traits/utl_variadic_traits.h
@@ -4,16 +4,16 @@
 
 #include "utl/base_preprocessor.h"
 #include "utl/memory/utl_uses_allocator.h"
-#include "utl/type_traits/assignment_traits/utl_is_copy_assignable.h"
-#include "utl/type_traits/assignment_traits/utl_is_move_assignable.h"
-#include "utl/type_traits/assignment_traits/utl_is_nothrow_copy_assignable.h"
-#include "utl/type_traits/assignment_traits/utl_is_nothrow_move_assignable.h"
-#include "utl/type_traits/constructor_traits/utl_is_copy_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_move_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_copy_constructible.h"
-#include "utl/type_traits/constructor_traits/utl_is_nothrow_move_constructible.h"
 #include "utl/type_traits/utl_add_lvalue_reference.h"
 #include "utl/type_traits/utl_add_rvalue_reference.h"
+#include "utl/type_traits/utl_is_copy_assignable.h"
+#include "utl/type_traits/utl_is_copy_constructible.h"
+#include "utl/type_traits/utl_is_move_assignable.h"
+#include "utl/type_traits/utl_is_move_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_copy_assignable.h"
+#include "utl/type_traits/utl_is_nothrow_copy_constructible.h"
+#include "utl/type_traits/utl_is_nothrow_move_assignable.h"
+#include "utl/type_traits/utl_is_nothrow_move_constructible.h"
 #include "utl/type_traits/utl_logical_traits.h"
 // reference_constructs_from_temporary
 #include "utl/type_traits/utl_std_traits.h"
